Fixed the issue with switching to classes and not holding of the gameboard borders
After I switched to ES6 and made instead of constructor function, classes, I also had an issue with not holding the gameboards borders. This was due to the use of array functions (like array.some() or for(let i of myArray). No I fixed it and also have shorten the code fro like 800 lines to 274. Yeah!